+++ src/components/form.js
@@ -1,4 +1,4 @@
-import {getRandomIntegerNumber, getRandomArrayItem} from '../utils.js';
+import {getRandomIntegerNumber, getRandomArrayItem} from '../utils/common.js';
 import {descriptions, pictures} from '../mock/form.js';
 import AbstractComponent from './abstract-component.js';
